refactor(rules): extract game rules into plain utility module
useRules had no reactive state, making the composable wrapper
unnecessary overhead. Moving the logic to a plain module removes
the Vue composition API dependency, simplifies imports across all
consumers, and makes the functions directly testable without
Pinia setup.
